+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Added `.gitattributes` file to standardize to LF line endings and renormalized codebase
 - Added invalid JSON graph test coverage for `saveGraphJSON` in `app/Controllers/Graph.hs`
 - Converted various JS files to TypeScript
+- Added static info existence test for `Database/Database.hs`
 
 ## [0.8.1] - 2026-08-10